+++ bed-utils/src/lib.rs
@@ -1,11 +1,2 @@
 pub mod bed;
 pub mod bam;
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        let result = 2 + 2;
-        assert_eq!(result, 4);
-    }
-}
