+++ snapatac2-core/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "project_a"
+name = "snapatac2-core"
 version = "0.1.0"
 edition = "2021"
 authors = ["Kai Zhang <kai@kzhang.org>"]
